# docs/api_spec.yaml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Nordend"
  contact:
    name: Amagi Nordend Team
    email: "stv-dev@amagi.com"
  description: |-
    <br>
    Nordend - A service which holds the record of the master metadata belonging to an Account, Channel, Delivery and Platform that Amagi offers.

    <br/>
    <hr/>
    <br/>

    ### Entities

    - **Account** : Account is a resource that is uniquely identifiable with AMGID and represents one billable customer.
    - **Channel** : Channel is a resource that represents any linear delivery specific workflow that is running within amagi ecosystem.
    - **Platform** : Platform is a resource that uniquely refers to any consumers of the deliverables that Amagi has to offer via linear or on-demand workflows.
    - **Delivery** : Delivery is an association entity which connects a account or a channel to one platform.

    <br/>
    <hr/>
    <br/>
    
    ### Entity Relationship

    | Entity | Association | Entity |
    |-|-|-|
    | Account | One to Many | Channel |
    | Channel | One to Many | Delivery |
    | Delivery | One to One | Platform |

    <br/>
    <hr/>
    <br/>
    
    ### API Usage

    #### Authentication

    - API Key Authentitcation is used to make any API calls.
    - API Key can be obtained by contacting the developer team .
    - API Key needs to be passed alongside the API authorization header for every API request.

    #### Response Codes

    | HTTP Verb | Allowed Responses |
    | - | - |
...

...
servers:
...
paths:
...
  /accounts/{amgid}:
    get:
      tags:
        - "Account Management"
      summary: "Get details of an account by amgid"
      description: |-
        This API allows the caller to get specific account resource.
      operationId: "getAccountByAmgId"
      parameters:
        - in: "path"
          name: "amgid"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        403:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unauthorized"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"


    delete:
      tags:
        - "Account Management"
      summary: "Delete a specific Account"
      description: |-
        This API allows the caller to delete an account resource by amgid.

        Note: 
          - This is an irrevocable action. Ensure you are absolutely sure before this is called.
          - This can be called, iff all the associated channel resources are deleted.

      parameters:
        - in: "path"
          name: "amgid"
          required: true
          schema:
            type: "string"
      operationId: "deleteAccount"
      responses:
        "200":
          description: "successful operation"
        403:
          description: "User forbidden from making the request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forbidden"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
  schemas:
...
    Unauthorized:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Unauthorized"
    Forbidden:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Forbidden"
...
    InternalServerError:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Internal Server Error"
...
    AccountResponse:
      type: "object"
      properties:
        amgid:
          type: "string"
          example: "AMG00001"
        name:
          type: "string"
          example: "Lhotse"
        logo:
          type: "string"
          example: "https://example.com/logo.png"
        description:
          type: "string"
          example: "Connect with Lhotse to broadcast your video to the world"
        domain:
          type: "string"
          example: "lhotse.now.amagi.tv"
...
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
security:
  - token: []